+++ src/main/java/annotation/User.java
@@ -1,3 +1,5 @@
+package annotation;
+
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
