+++ src/main/java/annotation/UserId.java
@@ -1,10 +1,14 @@
+package annotation;
+
 import lombok.AllArgsConstructor;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 
+@EqualsAndHashCode
 @AllArgsConstructor
-@ToString(includeFieldNames = false)
-public class UserName {
+@ToString
+public class UserId {
     @Getter
     public final String value;
 }
