+++ src/test/groovy/FixtureAddress.groovy
@@ -1,3 +1,4 @@
+import annotation.Address
 import groovy.transform.ToString
 
 @ToString
